+++ test/test.hs
@@ -2055,10 +2055,122 @@ tests = testGroup "hevm"
           }
         }
         |]
-      let sig = Just (Sig "checkval(uint256,uint256)" [AbiUIntType 256, AbiUIntType 256])
+      let sig = Just (Sig "checkval(uint8)" [AbiUIntType 8])
       (res, [Qed _]) <- withDefaultSolver $ \s ->
         checkAssert s defaultPanicCodes c sig [] defaultVeriOpts
       putStrLnM $ "successfully explored: " <> show (Expr.numBranches res) <> " paths"
+    ,
+    test "staticcall-check-orig" $ do
+      Just c <- solcRuntime "C"
+        [i|
+        // Target contract with a view function
+        contract Target {
+            function add(uint256 x, uint256 y) external pure returns (uint256) {
+              unchecked {
+                return x + y;
+              }
+            }
+        }
+
+        // Caller contract using staticcall
+        contract C {
+            function checkval(address inputAddr, uint256 x, uint256 y) public returns (uint256 result) {
+                Target t = new Target();
+                address realAddr = address(t);
+
+                bytes memory data = abi.encodeWithSignature("add(uint256,uint256)", x, y);
+                (bool success, bytes memory returnData) = realAddr.staticcall(data);
+                assert(success);
+
+                result = abi.decode(returnData, (uint256));
+                uint expected;
+                unchecked {
+                  expected = x + y;
+                }
+                assert(result == expected);
+            }
+        }
+        |]
+      let sig = Just (Sig "checkval(address,uint256,uint256)" [AbiAddressType, AbiUIntType 256, AbiUIntType 256])
+      (res, [Qed _]) <- withDefaultSolver $ \s -> checkAssert s defaultPanicCodes c sig [] defaultVeriOpts
+      putStrLnM $ "successfully explored: " <> show (Expr.numBranches res) <> " paths"
+    , test "staticcall-check-symbolic1" $ do
+      Just c <- solcRuntime "C"
+        [i|
+        contract C {
+            function checkval(address inputAddr, uint256 x, uint256 y) public returns (uint256 result) {
+                bytes memory data = abi.encodeWithSignature("add(uint256,uint256)", x, y);
+                (bool success, bytes memory returnData) = inputAddr.staticcall(data);
+                assert(success);
+            }
+        }
+        |]
+      let sig = Just (Sig "checkval(address,uint256,uint256)" [AbiAddressType, AbiUIntType 256, AbiUIntType 256])
+      (res, ret) <- withDefaultSolver $ \s -> checkAssert s defaultPanicCodes c sig [] defaultVeriOpts
+      putStrLnM $ "successfully explored: " <> show (Expr.numBranches res) <> " paths"
+      -- let cexesExt = map (snd . fromJust . extractCex) ret
+      -- putStrLnM $ "Cexes: \n" <> (unlines $ map ("-> " ++) (map show cexesExt))
+      let numCexes = sum $ map (fromEnum . isCex) ret
+      let numErrs = sum $ map (fromEnum . isError) ret
+      let numQeds = sum $ map (fromEnum . isQed) ret
+      assertEqualM "number of counterexamples" numCexes 2
+      assertEqualM "number of symbolic copy errors" numErrs 0
+      assertEqualM "number of qed-s" numQeds 0
+    , test "staticcall-check-symbolic2" $ do
+      Just c <- solcRuntime "C"
+        [i|
+        contract C {
+            function checkval(address inputAddr, uint256 x, uint256 y) public returns (uint256 result) {
+                bytes memory data = abi.encodeWithSignature("add(uint256,uint256)", x, y);
+                (bool success, bytes memory returnData) = inputAddr.staticcall(data);
+                assert(success);
+
+                result = abi.decode(returnData, (uint256));
+                uint expected;
+                unchecked {
+                  expected = x + y;
+                }
+                assert(result == expected);
+            }
+        }
+        |]
+      let sig = Just (Sig "checkval(address,uint256,uint256)" [AbiAddressType, AbiUIntType 256, AbiUIntType 256])
+      (res, ret) <- withDefaultSolver $ \s -> checkAssert s defaultPanicCodes c sig [] defaultVeriOpts
+      putStrLnM $ "successfully explored: " <> show (Expr.numBranches res) <> " paths"
+      let numCexes = sum $ map (fromEnum . isCex) ret
+      let numErrs = sum $ map (fromEnum . isError) ret
+      let numQeds = sum $ map (fromEnum . isQed) ret
+      assertEqualM "number of counterexamples" numCexes 2
+      assertEqualM "number of symbolic copy errors" numErrs 1
+      assertEqualM "number of qed-s" numQeds 0
+     , test "jump-symbolic" $ do
+      Just c <- solcRuntime "C"
+        [i|
+        // Target contract with a view function
+        contract Target {
+        }
+
+        // Caller contract using staticcall
+        contract C {
+            function checkval(address inputAddr, uint256 x, uint256 y) public returns (uint256 result) {
+                Target t = new Target();
+                address realAddr = address(t);
+
+                bytes memory data = abi.encodeWithSignature("add(uint256,uint256)", x, y);
+                (bool success, bytes memory returnData) = inputAddr.staticcall(data);
+                assert(success == true);
+            }
+        }
+        |]
+      let sig = Just (Sig "checkval(address,uint256,uint256)" [AbiAddressType, AbiUIntType 256, AbiUIntType 256])
+      (res, ret) <- withDefaultSolver $ \s -> checkAssert s defaultPanicCodes c sig [] defaultVeriOpts
+      putStrLnM $ "successfully explored: " <> show (Expr.numBranches res) <> " paths"
+      let numCexes = sum $ map (fromEnum . isCex) ret
+      let numErrs = sum $ map (fromEnum . isError) ret
+      let numQeds = sum $ map (fromEnum . isQed) ret
+      assertEqualM "number of counterexamples" numCexes 2
+      assertEqualM "number of symbolic copy errors" numErrs 0
+      assertEqualM "number of qed-s" numQeds 0
      ,
      test "opcode-mul-assoc" $ do
         Just c <- solcRuntime "MyContract"